tidy up code, remove some redundancies
